mp3 File Location Enhancement
The previous implementation was trying to solve the problem of
dynamically finding the plugin sub directory.  While this code
accomplishes that by adding a lot of extra complexity it still doesn't
solve the problem that if your OpenVBX installation is installed in a
sub directory the mp3 notification will not play.